feat(locale): footer hooked up to reload translation
