+++ packages/react/src/components/Footer/LocaleButton.js
@@ -1,5 +1,5 @@
 import { FOOTER_LOCALE_BUTTON } from '../../internal/FeatureFlags.js';
-import { featureFlag, ipcinfoCookie } from '@carbon/ibmdotcom-utilities';
+import { featureFlag } from '@carbon/ibmdotcom-utilities';
 
 import React, { useState } from 'react';
 import {
@@ -21,28 +21,15 @@ const { prefix } = settings;
  *
  * @private
  *
+ * @param {Function} selectItem method to handle selected item
+ *
  * @returns {object} JSX object
  */
-const LocaleButton = () => {
+const LocaleButton = ({ selectItem }) => {
   const [isOpen, setIsOpen] = useState(false);
 
   // retrieve list from local storage
   const list = JSON.parse(localStorage.getItem('countryList'));
-
-  /**
-   * method to handle when country/region has been selected from dropdown
-   *
-   * @param {object} item selected country/region
-   */
-  const selectItem = item => {
-    const stringLocale = item.selectedItem.locale[0][0];
-    const locale = stringLocale.split('-');
-    const objectLocale = {
-      cc: locale[1],
-      lc: locale[0],
-    };
-    ipcinfoCookie.set(objectLocale);
-  };
 
   /**
    *  method to sort list to be used by the dropdown
